+++ R/radf.R
@@ -44,7 +44,7 @@ psy_rule <- function(n) {
 #' @examples
 #' \donttest{
 #' # Simulate bubble processes
-#' dta <- cbind(sim_dgp1(n = 100), sim_dgp2(n = 100))
+#' dta <- data.frame(psy1 = sim_psy1(n = 100), psy2 = sim_psy2(n = 100))
 #'
 #' rfd <- radf(dta)
 #'
@@ -54,7 +54,7 @@ psy_rule <- function(n) {
 radf <- function(data, minw = psy_rule(data), lag = 0) {
 
   lst <- parse_data(data)
-  x <- as.matrix(lst$data)
+  x <- lst$data
 
   nc <- NCOL(x)