+++ R/wb_cv.R
@@ -77,7 +77,8 @@ wb_ <- function(data, minw, nboot, dist_rad) {
 #'
 #' \code{wb_cv} performs the Harvey et al. (2016) wild bootstrap re-sampling
 #' scheme, which is asymptotically robust to non-stationary volatility, to
-#' generate critical values for the recursive unit root tests.
+#' generate critical values for the recursive unit root tests. \code{wb_dist}
+#' computes the distribution.
 #'
 #' @inheritParams radf
 #' @inheritParams mc_cv
@@ -113,13 +114,16 @@ wb_ <- function(data, minw, nboot, dist_rad) {
 #' @examples
 #' \dontrun{
 #' # Simulate bubble processes
-#' dta <- data.frame("dg1" = sim_dgp1(n = 100), "dgp2" = sim_dgp2(n = 100))
+#' dta <- data.frame(psy1 = sim_psy1(n = 100), psy2 = sim_psy2(n = 100))
 #'
 #' # Default minimum window
 #' wb <- wb_cv(dta)
 #'
 #' # Change the minimum window and the number of bootstraps
 #' wb <- wb_cv(dta, nboot = 1500, minw = 20)
+#'
+#' # Simulate distribution
+#'wb_dist(dta)
 #' }
 wb_cv <- function(data, minw = psy_rule(data), nboot = 1000,
                   dist_rad = FALSE) {